Generalised the matching function to return the position and length of a matched substring.
In addition to

  accept :: RegExp a -> [a] -> Bool

there is now a function

  matchings :: RegExp a -> [a] -> [(Index,Int)]

which returns the starting index and the length of a matched
substring. Not only longest but all matches are returned.

There is an operator `(=~)` that specilises the type of the
`matchings` function to `String`s. It is useful in combination
with the language extension `OverloadedStrings` to play with
regular expressions as demonstrated by the following call:

  ghci> "a|abc" =~ "abc"
  [(1,1),(0,3)]