generalised Status type to use an arbitrary monoid.
The `RegExp` data type now has an additional type parameter `m`
that can be instantiated with a `Monoid` instance to determine
the matching policy.

Using `Any` checks whether the regexp matches a subword,
`Set Int` can be used to compute the positions of all matchings, and
`Min Int` can be used to find leftmost matchings.

The `Min` monoid is defined in the `Matcher` because it is not
predefined.

I have renamed `matchings` to `allMatchings`, added a function

    leftmostLongestMatching :: RegExp (Min Int) a -> [a] -> Maybe Matching

and redefined `(=~)` in terms of this new function.

On the go, I have refactored the `RE` data type which now also takes only
the monoid parameter `m` instead of an arbitrary type constructor `l` for
wrapping sub expressions. Now `RegExp` and `RE` are mutually recursive.