simple version.
I have extracted a specialised, restricted version of the approach
which is easier to understand but not as flexible. Instead of
supporting arbitrary monoids as matching policies, it only solves
the word problem. For the restricted interface, the simplified module
can replace the overloaded library in test programs (the interface is
almost identical).

The simplified version is slightly different from the first version
mainly because of the simplified RegExp type which is mutually
recursive with the RE type.

I have not compared performance extensively but the more general
overloaded version does not seem to be slower than the specialised one.
In the single test I have run, it is even a bit faster.