Reimplementation based on semirings.
I have rewritten the algorithm based on the semiring structure of
regular expressions. The matching function now maps a weighted
regular expression and a word to an arbitrary semiring element.
Weights are also semiring elements.

By using different semirings, different matching algorithms can be
implemented. I have implemented a functions that checks whether a
regexp matches a word using the Boolean semiring, another function
that counts the possibilities to match a regexp against a word using
the semiring of natural numbers, and another function that computes
all matchings of a regexp against subwords of a word using a more
complicated semiring based on sets of tuples.

After the change the algorithm is again shorter. The matcher fits
in 11 locs.