More efficient matching of subwords.
I have generalised the representation of regular expressions such that
symbols can have an associated weight that depends on the symbol *and*
its index in the original word.

Based on this I have re-added the matching algorithm which is restarted
on every character but also kept the one which matches the whole word.
The functions that compute all matchings or the leftmost longest matching
are now implemented more efficiently and I have added functions which
compute a list of all (or the leftmost longest) matching subword(s) of a
given word.