Implemented workaround for [GHC bug 4227]
Currently, GHC can SPECIALIZE functions only where they are defined.
The types `Leftmost`, `Longest`, and `LeftLong` are now defined in
separate modules to bring them into the scope of the matching functions.
This allows to specialize them for type defined in this package. Users,
however, must use the matching functions unspecialized for their own types.

With this change, the constructors of the matching types are no longer
exported.